project/pyproject.toml
[tool.poetry]
name = "ufficio urbanistica"
version = "0.1.0"
description = ""
authors = ["Massimiliano Moraca <info@massimilianomoraca.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
psycopg2-binary = "^2.9.3"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
django-jazzmin = "^2.5.0"
djangorestframework = "^3.13.1"
djangorestframework-gis = "^1.0"
django-tinymce = "^3.5.0"
Pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
